requirements.txt
PyQt5
python-can
